+++ CHANGES.rst
@@ -11,6 +11,7 @@ release.
 - Move GroupedQuerySelectField to WTForms-Alchemy
 - Move PhoneNumber to WTForms-Alchemy
 - Move WeekdaysField to WTForms-Alchemy
+- Move Unique validator to WTForms-Alchemy
 - Remove AJAXField dependency on SQLAlchemy-Utils